added comments to css, love quiz one and two (to explain duplication) and activity components
