+++ src/components/Activity/renderMusic.js
@@ -2,45 +2,37 @@ import axios from 'axios';
 import { useState, useEffect } from "react";
 import Container from '@mui/material/Container';
 import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
 import PageHeader from "../PageHeader";
 
 const RenderMusic = (props) => {
     const [music, setMusic] = useState("");
-    const [artistURL, setArtistURL] = useState("");
 
     const fetchMusic = async () => {
         const musicKey = "69f3c105afa838de9b436556ffd12348";
         let musicGenre = `${props.activityTopic}`;
 
         try {
-            const res = await axios.get("https://ws.audioscrobbler.com/2.0/?method=tag.gettopartists&tag=" + musicGenre + "&api_key=" + musicKey + "&format=json");
-            const index = Math.floor(Math.random() * res.data.topartists.artist.length)
-            setMusic(res.data.topartists.artist[index].name)
-            setArtistURL(res.data.topartists.artist[index].url)
+            const res = await axios.get("http://ws.audioscrobbler.com/2.0/?method=tag.gettopartists&tag=" + musicGenre + "&api_key=" + musicKey + "&format=json");
+            setMusic(res.data.topartists.artist[Math.floor(Math.random() * res.data.topartists.artist.length)].name)
         } catch (err) {
             console.log(err);
         }
     };
 
-    useEffect(() => {
-        fetchMusic();
-    }, []);
+useEffect(() => {
+    fetchMusic();
+}, [music]);
 
-    return (
-        <div>
-            <Container maxWidth="lg">
-                <Box sx={{ borderRadius: 2, marginTop: 2, marginBottom: 2, padding: 2 }}>
-                    <PageHeader title="SUGGESTED ACTIVITY" />
-                    <h4 className="quizInstructions">Discover a new artist together:</h4>
-                    <p className="quizInstructions">{music}</p>
-                    <div className="buttonDiv">
-                        <Button variant="contained" href={artistURL} target="_blank">Check them out!</Button>
-                    </div>
-                </Box>
-            </Container>
-        </div>
-    )
+return (
+    <div>
+        <Container maxWidth="lg">
+            <Box sx={{ bgcolor: '#e314d2', borderRadius: 2, marginTop: 2, marginBottom: 2, padding: 2 }}>
+                <PageHeader title="SUGGESTED ACTIVITY" />
+                <p className="quizInstructions">Discover a new artist together: {music}</p>
+            </Box>
+        </Container>
+    </div>
+)
 };
 
 export default RenderMusic
